+++ R/evaluation.R
@@ -10,7 +10,7 @@
 #'            data = NULL, values = NULL, values.audit = NULL, times = NULL,
 #'            x = NULL, n = NULL, N.units = NULL, N.items = NULL,
 #'            r.delta = 2.7, m.type = 'accounts', cs.a = 1, cs.b = 3, cs.mu = 0.5,
-#'            prior = FALSE)
+#'            prior = FALSE, predictive = FALSE)
 #'
 #' @param materiality   a numeric value between 0 and 1 specifying the performance materiality (maximum tolerable error) as a fraction of the total size of the population. If specified, the function also returns the conclusion of the analysis with respect to the performance materiality. The value is discarded when \code{direct}, \code{difference}, \code{quotient}, or \code{regression} method is chosen.
 #' @param min.precision a numeric value between 0 and 1 specifying the required minimum precision (upper bound minus most likely error) as a fraction of the total size of the population. If specified, the function also returns the conclusion of the analysis with respect to the required minimum precision.
@@ -31,6 +31,7 @@
 #' @param cs.b          if \code{method = "coxsnell"}, a numeric value specifying the \eqn{\beta} parameter of the prior distribution on the mean taint. Defaults to 3 as recommended by Cox and Snell (1979).
 #' @param cs.mu         if \code{method = "coxsnell"}, a numeric value between 0 and 1 specifying the mean of the prior distribution on the mean taint. Defaults to 0.5 as recommended by Cox and Snell (1979).
 #' @param prior         a logical specifying if a prior distribution must be used, or an object of class \code{jfaPrior} or \code{jfaPosterior} containing the prior distribution. Defaults to \code{FALSE} for frequentist planning. If \code{TRUE}, a minimal information prior is chosen by default. Chooses a conjugate gamma distribution for the Poisson likelihood, a conjugate beta distribution for the binomial likelihood, and a conjugate beta-binomial distribution for the hypergeometric likelihood.
+#' @param predictive    a logical specifying whether to include the prior predictive distribution in the \code{posterior} output. Requires the use of a prior distribution.
 #'
 #' @details This section lists the available options for the \code{methods} argument.
 #'
@@ -133,7 +134,7 @@ evaluation <- function(materiality = NULL, min.precision = NULL, method = "poiss
                        data = NULL, values = NULL, values.audit = NULL, times = NULL,
                        x = NULL, n = NULL, N.units = NULL, N.items = NULL,
                        r.delta = 2.7, m.type = "accounts", cs.a = 1, cs.b = 3, cs.mu = 0.5,
-                       prior = FALSE) {
+                       prior = FALSE, predictive = FALSE) {
   proper <- TRUE
   alternative <- match.arg(alternative)
   bayesian <- (class(prior) == "logical" && prior == TRUE) || class(prior) %in% c("jfaPrior", "jfaPosterior")
@@ -432,7 +433,7 @@ evaluation <- function(materiality = NULL, min.precision = NULL, method = "poiss
     } else {
       result[["prior"]] <- auditPrior(
         method = "sample", likelihood = method, N.units = result[["N.units"]],
-        materiality = result[["materiality"]], x = prior.x, n = prior.n
+        materiality = result[["materiality"]], x = prior.x, n = prior.n, predictive = predictive
       )
     }
   }
@@ -565,16 +566,16 @@ evaluation <- function(materiality = NULL, min.precision = NULL, method = "poiss
         }
         result[["posterior"]][["hypotheses"]]$odds.h1 <- result[["posterior"]][["hypotheses"]]$p.h1 / result[["posterior"]][["hypotheses"]]$p.h0
         result[["posterior"]][["hypotheses"]]$odds.h0 <- 1 / result[["posterior"]][["hypotheses"]]$odds.h1
-        # For improper priors we take the posterior odds as Bayes factor
+        # For improper priors we display the posterior odds
         result[["posterior"]][["hypotheses"]]$bf.h1 <- result[["posterior"]][["hypotheses"]]$odds.h1
-        if (proper) { # The prior is proper, so we divide by the prior odds
+        if (proper) { # The prior is proper, so we divide by the prior odds to get a Bayes factor
           result[["posterior"]][["hypotheses"]]$bf.h1 <- result[["posterior"]][["hypotheses"]]$bf.h1 / result[["prior"]][["hypotheses"]]$odds.h1
         }
         result[["posterior"]][["hypotheses"]]$bf.h0 <- 1 / result[["posterior"]][["hypotheses"]]$bf.h1
       }
     }
     # Create the posterior predictive section
-    if (method != "hypergeometric" && !is.null(result[["N.units"]])) {
+    if (predictive && method != "hypergeometric" && !is.null(result[["N.units"]])) {
       result[["posterior"]][["predictive"]] <- list()
       result[["posterior"]][["predictive"]]$predictive <- switch(method,
         "poisson" = paste0("Negative-binomial(r = ", round(result[["posterior"]][["description"]]$alpha, 3), ", p = ", round(1 / (1 + result[["posterior"]][["description"]]$beta), 3), ")"),