+++ NEWS.md
@@ -1,3 +1,9 @@
+# jfa 0.6.2
+
+**Minor changes**
+
+- Added argument `predictive` to the `auditPrior()`, `planning()` and `evaluation()` functions that provide control over whether to include the prior predictive distribution in the output. This may take a long time for the beta distribution.
+
 # jfa 0.6.1
 
 **New features**